Warn (and error) on undefined functions
a) Warn on undefined tags. First time I am using a
define-compiler-macro, so hopefully this is not an overkill use of
this mechanism. I might use compiler macros to optimize certain things
in the future though.

b) I noticed I wasn't giving the function name on error, this was
because of a recent refactoring. So I fixed that.